refactor: move reusable logic and enum registrations
